[WIP] Release 0.15.2
--- TODO

 - Set final release date in README.md
 - Generate final tarballs after date is set

--- GitHub release announcement text:

Another bugfix release without API- or ABI-changes, mainly to
fix more known 0.15.0 and 0.15.1 regressions, but as a bonus
it is now possible to build the DirectWrite system fontprovider
for Microsoft's Universal Windows Platform (UWP).
There remain two known unfixed 0.15.x regressions for DirectWrite
occuring when either some but not all used variants of a font are
attached or on the rare occurence of the full name of a font equaling
the (non full) family name of another font. The former also affects
Core Text since libass 0.13.0. Fontconfig is unaffected.

Also please note that our IRC channel #libass moved from freenode.net
to libera.chat and all presence on freenode.net was fully retired.

--- NOTE

  If generated with automake 1.16.4+ the tarballs will contain
  the README.md and compare/README.md files. Previously those were
  not included (but a plain README would have been)

--- DONE

 - Update the versions in ass.h, configure.ac, README.md and Makefile.am
 - Create tarballs with autoconf 2.71
   (gen. on amd64 glibc-Linux, but still wrong README-date,
    so not for upload but only comparison to final tarballs)
 - distcheck with (autoconf-version; automake-version):
   Locally:
     * amd64 glibc-Linux gcc   (2.69; 1.16.1) and (2.71; 1.16.4)
     * i586  musl-Linux  gcc   (2.71; 1.16.3)
     * amd64 Illumos     gcc   (2.71; 1.16.4)
   GHA ( https://github.com/TheOneric/libass/runs/3356548731 )
     * amd64 musl-Linux  gcc   (2.71; 1.16.3)
     * amd64 MacOS-10.15 clang (2.71; 1.16.4)
     * i686  Windows     gcc   (2.71; 1.16.3)
 - regression tests on amd64 Illumos
 - compile from _pregenerated_ tarball and regression test
   on Android+Termux (bionic-Linux, armv7-eabi)
   (after upgrading my packages, autoreconf is bugged so no distcheck;
    but that's definitely not a libass but Termux packaging issue)